Adding some tests of the substring method, motivated by my recent Chapel scripting work.  Questions that I wasn't sure of the answer two were:
1) what should happen if a substring index is out of
   bounds?

2) what should happen if a substring range is
   degenerate?

Looks like for (1) we ignore the out-of-bounds indices,
which I wasn't sure was appropriate, so I filed a future.
For (2) looks like things are working as I'd expect --
the empty string is returned.  However, in one case, I'm
getting an "out of memory" error, making me wonder if
we're not actually implementing (2) correctly, so filed
a bug against that case.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12243 3a8e244f-b0f2-452b-bcba-4c88e055c3ca