Change the functions to add and delete roots for GC into C macros. They are extremely performance critical and are way too slow as functions.
Add PRIMITIVE_GC_CLEANUP and a call to it at the end of _chpl_main.  It is a
runtime function that frees the GC heaps before the main finishes. Although this
isn't necessary since the program is about to terminate, it makes for cleaner
valgrind reports and --memstat reports.

Fix the valgrind allocateManyObjects regression. On a machine where an enum is
smaller than a void*, the last bits of the _padding_for_copy_collection_use
field were not set when the class id field it is in a union with was set. In the
code for setting the class id field, first set the padding field to NULL.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12294 3a8e244f-b0f2-452b-bcba-4c88e055c3ca