Tightened up handling of ".type" to mark these expressions as types so that they cannot, for example, be passed to functions that expect values.  David noticed this issue a while back.  This change fixes an issue with codes like "writeln(x.type)" where x is a variable; this code is now an error, but used to work, in a way.
Improved function printing for error messages to mark variable-length
arguments with "..."

Added support for indexing into tuple types (handled via preFold) to
get the type in the tuple.  This supports indexing into a type of a
variable-length argument.

Added support for determining the type of an array via ".type".

Defined writeThis method on the internal _stack type, removing
overloaded writeln.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12299 3a8e244f-b0f2-452b-bcba-4c88e055c3ca