Marked methods that return type fields as type functions.
Removed code in function resolution pre-folding in which a member
access of an instantiated type would be replaced by the instantiated
type.  This is no longer necessary because type members are available
via type methods.  This removes the section of code that was
associated with parameter members (yesterday's change) in total.  This
is especially necessary for arrays of arrays in that the type field
will need to be changed into a value to store the runtime array type.

Added code to post-fold the result of type functions (think inlining)
if the type is not an array type (in which case it will have to be
changed into a runtime array type later).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12317 3a8e244f-b0f2-452b-bcba-4c88e055c3ca