Followed up on the last Makefile-related check-in (which seems to be working well), by pushing object files and library files down into a compiler-specific subdirectory.
Thus, the current answers to "where do things live" is:

compiler-related object files and "make depend" files:
  compiler/<subdir>/$CHPL_PLATFORM/$CHPL_COMPILER/*.o

compiler:
  bin/$CHPL_PLATFORM/

runtime-related object files and "make depend" files:
  runtime/<subdir>/$CHPL_PLATFORM/$CHPL_COMPILER/threads-$CHPL_THREADS/comm-$CHPL_COMM/

runtime libraries and main.o:
  lib/$CHPL_PLATFORM/$CHPL_COMPILER/threads-$CHPL_THREADS/comm-$CHPL_COMM/


In the asbove, CHPL_PLATFORM and CHPL_COMPILER refer to the
host options for the compiler and for the target options for
the runtime.

The goal of this is to allow a user to switch between compilers
without having to clean out their directory.  For example, one can
switch between gcc and intel, or between the three back-end options on
the XT without doing a make clean, or needing to worry about using
object files from different compilers.

Note that I made the choice to leave the compiler at
bin/$CHPL_PLATFORM rather than pushing it down another level.  My
hypothesis is that the compiler shouldn't care which compiler it
was built with (and hopefully, neither should the user).  The
runtime libraries do, since they will get linked by a particular
compiler of course.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12371 3a8e244f-b0f2-452b-bcba-4c88e055c3ca