Changed the HPL benchmark to factor A = [Asquare | b], where the linear system is Asquare*x = b.  This eliminates the forward solve step after the factorization and it matches the reference implementation for HPL.  Also, it is no longer necessary to return the pivot vector from the factorization routine for use in the solve routine.
The matrix A = [Asquare | b] is passed to the initialization routine.
Rather than initialize A with one call to fillRandom, I modified the
initialization routine to expect this (n x n+1) matrix, initializing
Asquare first and then b.  This matches what was previously done when
A and b were stored separately.  (Then I could test against the previous
version and have the same linear system.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12373 3a8e244f-b0f2-452b-bcba-4c88e055c3ca