Folded what little constant checking was done in the post-resolution semantic checks pass into function resolution by
  - marking arguments with blank or const intent as constant only if
    they are not references, arrays, or domains

  - flagging the creation of references of constants as an error

  - flagging assignment to constants or parameters as errors

This should eliminate the need for the isExprTemp flag as marking the
temporary as constant should achieve the same results.

Fixed the _set_field function which assigned to a constant argument in
error but which was previously not caught.  This lead to the
difficulty with _promotionType described below.

Fixed setIndices in associative domains to not assign to "this" which
is constant.  This assignment is just unnecessary anyway.

Updated a few error messages to use the errors in function resolution
instead of semantic checks.  In some cases, these are better; in
others, they are just as good.

Removed code to initialize the "_promotionType" field of a class from
being partially generated.  This should fix the baseline regression
regarding the array of array of array test case.  The above changes
made this regress even with optimizations which is why these changes
are coupled together.

Added another array of array assignment test and an array of array of
array assignment test.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12420 3a8e244f-b0f2-452b-bcba-4c88e055c3ca