Made the elimination of references to array and domain wrappers as well as iterator classes complete in cullOverReferences.  Previously this was just done to handle functions with valid var pragmas.  Now that iterating over arrays of arrays yields an array wrapper, we need to either eliminate such references everywhere or nowhere.  It appears that eliminating them everywhere is possible.  To eliminate them nowhere would require putting the reference returned by "valid var" functions on the heap.  This would make garbage collection difficult.
Added the ability to return a "this" in a class's initialize function.
When something is returned by the initialize function of a class, the
constructor returns whatever it returns.  This is essential for domain
wrapper records.  Since references are stripped from domain wrapper
records, the modification to "this" in the initialize function was not
reflected in the correct wrapper.  Reworked the domain wrapper's
initialize method to take advantage of this feature.

Modified compute_sym_uses to mark all array actuals as use and def
since arrays are passed by reference.

Disabled scalar replacement of array and domain wrappers.  To be
reenabled separately in order to make this change easier.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12444 3a8e244f-b0f2-452b-bcba-4c88e055c3ca