Added partial bounds checking on partially bounded ranges used for slicing.  Thus given the array
  var A: [1..3] int;

the slice

  A(4..)

is now an error rather than an empty array.

Updated the blockLU code to do a check before computing "the k-th
elimination step" as in the new code (marked by +):

      //   store multipliers...
+     if k+1 <= A1.domain.high(1) then
        A1[k+1.., k..k] /= pivot;

Question.  Should we do these bounds checking or should slicing
silently result in the empty array?  If it does result in the empty
array then should A(4..5) silently result in the empty array too?


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12490 3a8e244f-b0f2-452b-bcba-4c88e055c3ca