Added a warning to inform the user that a function does not return a value (or halt) on all control paths.
Currently, the identification of a halt is done by matching the name
of the function called with "halt" as opposed to identifying functions
that exit on all control paths.

Wrappers no longer inherit the pragmas of their functions (since
inheriting "specified return type" was a problem).  This pragma is
used to inform this check that the initialization of the returned
symbol should not count in determining whether the returned symbol has
been assigned.

Made the "gcd" test issue this warning.  Updated the
test_for_param_return test so that it halts if a return is not taken.

Removed three futures.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12492 3a8e244f-b0f2-452b-bcba-4c88e055c3ca