Implemented the threads-mta version of the wait, locking, and signaling functions.  In order to be efficient on the MTA, the signaling functions need to also unlock, so ChapelBase.chpl was changed to reflect this, as was the pthreads implementation.  The threads-none version did not need to be changed, since it doesn't need to lock or unlock anything.
Note that the is_full primitive is not yet implemented on the MTA, since the
same primitive is currently used for both simple and nonsimple sync variables.
Consequently, programs that use write or writeln will abort at run time on the
MTA, since these functions call the is_full primitive.  This will be fixed in
a later check in.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12494 3a8e244f-b0f2-452b-bcba-4c88e055c3ca