Made all strings used in the compiler canonicalized strings.  Thus they can be compared for equality by address and they are all stored in a hash table that is deleted when compilation completes.
Removed the 1024 length restriction on strings as well as the
automatic truncation.  We may want to truncate all strings that we are
code generating (due to the PGI bug).

Removed old string utility functions in favor of astr (which
concatenates strings or turns a single string into a canonical string)
and istr (which turns an integer into a canonical string).  Added
asubstr to return a prefix substring of a string in canonicalized
form.

Changed many char* pointers into const char* pointers.

Removed a future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12515 3a8e244f-b0f2-452b-bcba-4c88e055c3ca