Added a compiler check for recursive type aliases, an error.
Fixed three .goods and removed three futures.  These tests are all in
error because type aliases, like types in general, can be used before
they are defined.  This makes two of these tests have recursive type
aliases.  The other one, which has the lines

    type string = complex(128);
    type complex = bool;

is a problem because the type alias string is defined in terms of the
type alias complex to form the illegal type "bool(128)".

Added a test.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12555 3a8e244f-b0f2-452b-bcba-4c88e055c3ca