Added an internal function _cond_invalid to check an expression to determine if it is valid as a condition either of a while statement or of an if statement or expression.  Used this function in the implementation of && and || to test for an illegal promotion.
Used the _cond_test function in select when clauses to ensure that the
result of the == function is valid as a condition either of a while
statement or of an if statement or expression.

Updated the future messages of two tests.  I'm now happy with the
behaviour of the compiler as regards these two tests, but I did not
remove the futures pending further discussion.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12615 3a8e244f-b0f2-452b-bcba-4c88e055c3ca