Implemented casts from complex to imaginary, real, and integral types. Note that the cast from complex to complex is already handled by the casts that were implemented to complex.
These casts were previously working by going through strings because
of a coercion from complex to string and then a cast of the string!!!
That failed for imaginary because of the strict error checking.  I'm
not sure we ever tested this in any event.

Added a test case

  var i: imag;
  i = (i + 1):imag;

I believe the output

  0.0i

is correct because (i+1) is 1.0+0.0i, a complex value, and casting
that to an imaginary takes the imaginary component which remains 0.0i.
Does anybody disagree with that?


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12617 3a8e244f-b0f2-452b-bcba-4c88e055c3ca