Added support for "point of instantiation" semantics and removed a future.  This says that an instantiated function uses the call for which it was instantiated as its definition point in order to determine the visible functions when resolving calls in its body. This is implemented via SymScope* Symbol::instantiationPoint which points to the scope of the call for which the function was instantiated.  Following C++, if a function is instantiated by more than one call, the an arbitrary call is chosen as the instantiation point.
Unmarked functions with class arguments as global.  All class methods
are still marked as global in order to support dynamic dispatch.
Without multiple dispatch, it is no longer necessary to mark any
function with a class argument as global.

Updated three .goods which changed due to the visibility changes of
functions with class arguments.  Added a test of a nested function
with class arguments that refers to an outer variable.

Updated the spec.

Note that I inserted a hack in function resolution to handle an odd
case where the compiler-generated _cast function for records would
call the constructor of a record from within it.  This could probably
work with a function resolution that works out-of-order which I think
should probably be bumped up in priority next year.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12721 3a8e244f-b0f2-452b-bcba-4c88e055c3ca