Added ability to implement machine-specific behaviors by defining a chpl_md.h file (md == machine dependent) to the runtime/include directories.  The default one does nothing; a platform may override this default in its subdirectories by adding its own chpl_md.h file. This file should only be #included using angle bracket includes (i.e., #include <chpl_md.h>), and hopefully will only ever need to be included in stdchpl.h
Used this capability to implement the 32-bit math
routines for the Suns -- added these to chplmath.h,
but surrounded by an #ifdef that can be defined
within a platform's chpl_md.h file.  This may be
useful on the MTA as well, depending on how many
32-bit routines it supports (it doesn't support
sqrtf... I've got an inquiry out to the team).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12747 3a8e244f-b0f2-452b-bcba-4c88e055c3ca