Reworked this future to report the number of bits in the result of the sqrt function called on int(8), int(16), int(32), int(64), real(32), and real(64).  Currently, when called on int(32) and int(64), the result is a real(32).
The proposed change is to remove coercions from int(32) and int(64) to
real(32), but leave the others.  I think this makes sense because
real(64) is our default real size and it is odd that a default int
would lose precision and go to something less than the default real
when coercing it to the default real would not lose precision.

As written, this test didn't show this because the print routine only
prints out a few digits, not enough to show a difference when taking
the square root of 5.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12755 3a8e244f-b0f2-452b-bcba-4c88e055c3ca