Removed implicit conversions from any int or uint type to real(32) or complex(64).  There is still a coercion to real(64) and the coercion from int(64) to real(64) still loses precision.
This is slightly different than what I originally had in mind because
int(8), uint(8), int(16), uint(16) cannot coerce to real(32) or
complex(64) as I originally intended.  This would result in an
ambiguity on an int(8) + int(8), for example, between int(32)+int(32)
vs. real(32)+real(32) since neither is determined to be better.

Updated the spec and added a rationale.

Removed a future.

Can some casts be removed?  From madness?


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12762 3a8e244f-b0f2-452b-bcba-4c88e055c3ca