Added support for a --no-checks flag that disables runtime checks (currently --no-nil-checks and --no-bounds-checks).
Made the --no-nil-checks and --no-bounds-checks flags into the
variety of flag that can have its "no-" removed, so that one
can throw the --no-checks flag and then selectively turn
certain things back on.  This isn't so useful today, but as
we continue adding runtime semantic checks, it should become
more useful.

Added a new flag type ('n') in order to support these --no-
flags.  In the previous ones (like --no-cg-cpp-lines) using
the --no- version set the flag to false; with this new one,
the --no- version sets the flag to true.

Added a comment to arg.cpp and driver.cpp listing the flag
types and what they mean -- this has never really been
documented.  This is a step in the right direction.

Updated the man page and help test messages.

Renamed the --no-cg-cpp-lines flag into --no-cpp-lines.
The "cg" stood for codegen, but seemed increasingly
extraneous and confusing.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12764 3a8e244f-b0f2-452b-bcba-4c88e055c3ca