Stop using dynamic dispatch on methods with no parenthesis.
Move the "outer" pointer that nested classes have to the end of their field
lists. Previously it was at the beginning of the list. This was required because
the beginning of a derived class needs to have exactly the same shape as the
class it extends so that inherited fields can be accessed through a C pointer
cast to the base type.

Updated two tests to use methods with empty parenthesis instead of no parens so
they continue to get dynamic dispatch.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12805 3a8e244f-b0f2-452b-bcba-4c88e055c3ca