Re-enabling support for sparse domains and arrays within classes using a somewhat hacky technique -- used the setIndices/getIndices interface to copy value information from one domain to the other (this is also what made me realize that sparse domain assignment had never been implemented).
The more proper way to do this is either to create
a runtime type for the sparse domain that includes
its bounding box, or to generalize domain assignment
for domains of matching type to allow the domain
implementation to copy arbitrary value information
from one domain to the other.  There's also a little
bit of a semantic question tucked into this issue:
should assigning sparse domains assign their bounding
boxes as well?  Should a sparse domain's bounding
box be allowed to be modified at all?


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12822 3a8e244f-b0f2-452b-bcba-4c88e055c3ca