Adding an opaque domain/array test to the release and two futures.  The release example creates a small sample opaque domain/array representing a father/son relationship between three people. It then generates a random graph with weights on edges and vertices.
The futures show that a seg fault is possible in
opaque domain/array codes once they grow beyond
a certain size (possibly due to the associative
domain/array implementation).  The other gives
an example of an error message that could be
improved -- accessing an opaque array using nil
results in an out-of-bounds error instead of
an error message that refers to the nil value
(which should be clearer given that it's likely
to be a common case).

Still need to update the READMEs and Makefiles,
but will finish my other examples first.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12863 3a8e244f-b0f2-452b-bcba-4c88e055c3ca