Added comments to the purge() methods indicating that they are weird, and pointing to this and my previous CVSLOG message as documentation for why.
And to finish that documented: remembered why creating
standalone functions to try to implement purge()
doesn't work -- one needs to know the eltType of the
SingleLocaleAssociativeArray, but since the domain
stores a list of BaseArrays (and we're a single-dispatch
language), there's no way to write a standalone function
that resolves and also reveals eltType.  One would need
to dynamically cast from a BaseArray to a
SingleLocaleAssociativeArray, but there's no way to do
this without already knowing what the eltType type is.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12914 3a8e244f-b0f2-452b-bcba-4c88e055c3ca