Adding two more test matrices for BlockLU.  Both of these matrices have a condition number of approximately 50.  (The current matrix input for BlockLU is a magic square and has a condition number of 10^(17).  That is very ill-conditioned and may lead to differing results depending on platform and compiler.)
These matrices were generated in Matlab with the following commands:

A = gallery('randcolu',10);
save randcolu10.dat A -ASCII;

A = gallery('rando',10);
save rando10.dat A -ASCII;

I then had to add insert a line at the start of each file indicating
the number of rows and columns in the matrix.  (The BlockLU routine
expects to read in the number of rows and columns of the matrix before
reading in the matrix values.)

To learn about generating other matrices in Matlab, type "help gallery."
There are all kinds of matrices that can be generated in Matlab, including
sparse ones.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12924 3a8e244f-b0f2-452b-bcba-4c88e055c3ca