Removed UnresolvedSymbol subclass.
Added char* SymExpr::unresolved to handle symbols that are unresolved.

This changes decreases the memory footprint by a little over 2MB when
compiling Jacobi.  I believe it will help with simplifications along
the way.

For symbols that are unresolved before scope resolution, the SymExpr
has a null var field and a set unresolved field.  For function calls
that are unresolved before function resolution, the SymExpr in the
baseExpr of the CallExpr has a null var field and a set unresolved
field.  After function resolution, unresolved is no longer used.

Verification ensures that either unresolved or var is set but not
both.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12945 3a8e244f-b0f2-452b-bcba-4c88e055c3ca