Reimplemented insertWideReferences so that every reference in the entire program is changed into a wide reference.  Not surprisingly, this adds significant serial overhead.  This overhead can be disabled via the --local flag (the default with CHPL_COMM=none; see below).
Changed how wide references are supported.  It used to be the case
that wide references required their own primitives in order to
remotely get or put data.  Now, the same primitives that are used for
normal references are also used for wide references.  This should make
it much easier to optimize the code because changing a wide reference
into a reference will be as easy as changing the type of that
variable.

This handling requires more "normalized" code where temporaries are
inserted to capture references consistently.  Fixed code in the
copyCollect pass in which dereferencing was not always done at the
statement level.

Improved line number reporting for cobegin and begin functions.

Fixed a bug in which an infinite recursion could result when
hierarchically order records and unions for code generation.  This
showed up when the "outer" reference of a nested class was changed
into a wide reference record.

Added more value temporaries for primitives whose arguments are
references.  This is currently special-cased for about 12 primitives,
but should really be done for most.  Added a comment.

Unified the bundling of arguments for interfacing with runtime
functions.  This used to be handle separately for begin- and
cobegin-statements and for on-statements.  Now all three constructs
are handled with the same code at the same time.

Added macros to chplcomm.h for setting up wide references and calling
the runtime communication interface functions more cleanly via wide
references.

Changed --local into --[no-]local so that it can be turned off or on
at the command line.  Made the default key off of the CHPL_COMM
environment variable such that if CHPL_COMM is not set or is "none,"
then the default is --local; otherwise the default is --no-local.

Implemented comm-none runtime functions using memcpy.  This allows
--no-local to be used with CHPL_COMM=none.

Added --no-local to the set of flags tested via OPT_TESTING_FLAGS.

Removed a future related to writing an immediate value to a remote
record.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12977 3a8e244f-b0f2-452b-bcba-4c88e055c3ca