Fixed a bug in which wide references were passed to external functions that expected references (via INOUT intent).  These wide references are now assumed to be local and passed.  Next step, maybe: check to make sure the wide reference is local and halt if it is not.
Fixed a bug in remotely assigning nil to a record field via
_COMM_WIDE_PUT_OFF.  Unified writing an immediate to a record field
with other writes, much like is done for remotely assigning a
variable.  This removes the need for _COMM_WIDE_PUT_OFF_IMM.

Fixed a bug in which the type passed to _COMM_WIDE_PUT_OFF was taken
to be the type of the rhs rather than the type of the field.  This
caused a problem with assigning nil because the type of nil is
_nilType or void.  When optimizing with the C compiler using -O3, we
see the warning: "dereferencing type-punned pointer will break
strict-aliasing rules" (because we end up dereferencing void*) and get
incorrect semantics in the compiled code.

This should fix the UTS regression on GASnet.

Removed kludge in CallExpr::codegen related to the old implementation
of complex numbers.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12996 3a8e244f-b0f2-452b-bcba-4c88e055c3ca