This puzzle
  writeln([[1..3] 1..3))

is understood by reading the expressions inside-out.  The inner expression
is the forall-expression

  [1..3] 1..3

that is equivalent to the longer

  [i in 1..3] 1..3

and the even longer

  forall i in 1..3 do 1..3

This expression is equivalent to an iterator that returns the range 1..3
three times

  1..3 1..3 1..3

This forall-expression then results in a promotion of the [) operator and
is equivalent to an iterator that returns the expression

  [1..3) [1..3) [1..3)

or

  1..2 1..2 1..2

If captured into a variable, the result is a three-element array where
each element is the range 1..2.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13011 3a8e244f-b0f2-452b-bcba-4c88e055c3ca