Added initial support for the --gdb flag when using GASNet/multilocale execution.  In particular, on Cray Linux machines when using GASNET_SPAWNFN L, you should get an xterm launched for each locale that is running gdb such that a "run" command will cause it to start executing.
This support could be made more general and convenient
by: removing the absolute path to "xterm"; providing
some sort of optinal capability to automatically run
the gdb sessions and exit them when the program exits;
setting automatic breakpoints like those we provide
with the CHPL_COMM=none implementation, but this is a
useful enough starting point to be worth checking in.

Note that the setting of GASNET_ROUTE_OUTPUT will
alter the effects of whether an individual locale's
output will go to its gdb window or back to the main
console.  I haven't played around with this enough
yet to be able to describe what happens and why (I
believe that using GASNET_ROUTE_OUTPUT 0 will cause
output to go to the gdb window and that unsetenv-ing
it will cause it to go to the console).  This should
become a moot point once we have Chapel output being
communicated from locales to a host locale anyway.

To implement this change, I ammended the
_chpl_comm_init() call to take a third parameter
"runInGDB" that indicates whether or not the
execution should take place in GDB.  For CHPL_COMM=

none: we move the previous code that was used to
      launch the program in GDB into the chpl_cfg_comm.c
      file

gasnet: we set an environment var that indicates to
        the GASNet implementation that it should add
        the xterm and gdb arguments to the child
        processes it spawned.  This involved modifying
        the GASNet code because GASNet's initialization
        procedure wants to have pretty tight control
        over the argc/argv that's used for child
        processes.  Another way to do this might be to
        attach to the processes with debuggers after
        the fact, but I wasn't clear how to get the
        process numbers.  The GASNet team might have
        some advice here if we wanted to pursue this
        later.  I used an environment variable to
        avoid changing the GASNet interface.  Also note
        that I only changed the UDP implementation (I
        think).

armci: I simply threw an error if someone tries to use
       --gdb with armci for the time being.  I can't
       get armci to compile on our desktops due to the
       reliance on MPI which we don't seem to have
       installed anymore.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13041 3a8e244f-b0f2-452b-bcba-4c88e055c3ca