Implemented casts of wide classes using a macro so that we can do two assignments (one for the locale, one for the address) and avoid the "dereferencing type-punned pointer will break strict-aliasing rules" warning.
This change requires more consistent normalization of primitive casts.
Refactored code in reference propagation to insert an extra temporary.
Also, in insertWideReferences, inserted an extra temporary to capture
casts if the type of the cast does not match the type of the move
target; this allows us to separate the put from the wide cast.

This checkin fixes the GASnet issues distinct from --no-local
encountered when compiling with -O.  The current failures related to
the wide class change are limited to the following tests:

          classes/deitz/dispatch/test_dc1
          classes/deitz/dispatch/test_dd1
          classes/diten/allocateManyObjects


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13069 3a8e244f-b0f2-452b-bcba-4c88e055c3ca