This check-in results in a medium-sized change to the start_test script: it eliminates the -onetest and -startdir flags and supports an arbitrary number of command-line arguments that will represent directories (determined by -d and -x tests) and files (determined by -r and not being a directory) to test.
As before, directories are tested recursively by default
unless the -norecurse flag is used.

If no directories or files are provided, the directory
list is assumed to be "." (which is to say, using
"start_test" with traditional flags other than -startdir
and -onetest will work as before).

Irrespective of the order they are provided on the
command line, all files are tested before any directories.
The files and directories themselves are tested in the
order specified.

This check-in also updates any scripts that I could find
that use the -onetest or -startdir flags with start_test.
It's likely that I've missed some, so check for any that
you particularly like.  Also, if you own any cron jobs,
please check to make sure that they don't rely on these
flags (or check tomorrow to ensure that all your cron
jobs have run and returned the right number of successes).

This change does not include the change that we discussed
after today's meeting to allow start_test to be run from
arbitrary directories.  That will come in a distinct checkin.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13078 3a8e244f-b0f2-452b-bcba-4c88e055c3ca