Following discussion after today's meeting, rewrote start_test script so that it can be run anywhere. If no directories or files are listed as arguments to test, it will start at "." and work downwards looking for tests *unless* it is in $CHPL_HOME, in which case it will start in ./test.
The main impact of this change on most users is
that $CHPL_HOME will need to be set before running
start_test.

In addition, start_test has moved from the test/
directory to the util/ directory, so you'll want
that in your path if it isn't already.

Let me know if you find any problems of course --
the test scripts are difficult to test themselves.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13083 3a8e244f-b0f2-452b-bcba-4c88e055c3ca