Updated nightly script to make it more sensitive to host/target platform, compiler, threads, and comm settings in the names of its files and the mail that it sends out.
In doing so, removed the -gasnet flag and made it
trigger off of the CHPL_COMM environment setting
instead.

This will make it look like we have no ongoing
regressions and many new regressions the first
time each platform runs this new script due to
the new filenames that will be generated, but
on the plus side, will allow us to test intel
and get unique filenames and subject lines.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13105 3a8e244f-b0f2-452b-bcba-4c88e055c3ca