Merged the parallel nightly testing script (nightly-par) into the normal nightly script.  To get a parallel run, use the -parnodefile option, providing a nodefile such as one would supply to paratest.server.pl.  I think this currently needs to be an absolute path, but am not sure why that would be the case.
This change makes the nightly parallel testing script
much more robust.  In particular, the previous script
had hard-coded the nodefile into the script, making it
pretty useless if you didn't have my CHAP09 file in
your testing directory.  That said, this is still likely
to be fragile, due to the limited amount of testing that
it has received so far.

This merge has the benefit of adding much of the recent
cleanups in the nightly script into the parallel nightly
script (which was a branch of the nightly script that
had not been kept up-to-date).

Removed nightly-par.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13113 3a8e244f-b0f2-452b-bcba-4c88e055c3ca