A few tests that may shed some light on the previous future that I checked in.  These show that for some reason an array of sync class variables, even not in a class, requires a 0-argument constructor.  It doesn't seem like it should.
Also strange is that if the class only has simple
members (an int), it works.  But that if the class
has a range member (or more generally a generic
member?) it doesn't work.

Is type "range" with no decoration generic?  I assumed
that it was a range with default int element type.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13149 3a8e244f-b0f2-452b-bcba-4c88e055c3ca