Added initial support for using global variables in a multilocale program.  This greatly improves support for multilocale programs. Details are below.
There are several expected CHPL_COMM=gasnet and --no-local
regressions, but I wanted to check in this code before our in-town
meetings.  The following tests regress with CHPL_COMM=gasnet or
--no-local:

  compiler output error - incorrect line number
    classes/bradc/overloadMethods/v2NoReturnValueButType

  compiler output error - recursive function return type
    classes/deitz/tree
    functions/deitz/iterators/tree
    parallel/cobegin/bradc/varsEscape-workaround
    release/examples/iterators

  program output error - undiagnosed nil dereference
    arrays/deitz/test_array_dot_type
    arrays/deitz/part2/test_array_init

  program output error - sizeof primitive of heap class
    trivial/deitz/test_type_sizes

  program output error - unions and references
    classes/bradc/writeunion-workaround
    classes/bradc/writeunion
    classes/bradc/unions/assign
    classes/bradc/unions/assign1a
    classes/bradc/unions/stringunion
    classes/bradc/unions/twounions
    classes/bradc/unions/writeunion
    classes/deitz/test_union1
    classes/deitz/test_union2
    classes/deitz/test_union3
    classes/deitz/test_union4
    classes/deitz/test_union_method1
    classes/deitz/test_union_method2
    classes/deitz/test_union_method3

Extended the comm interface for the runtime with two functions and
implemented these functions for none and gasnet.  They are described
as such:

  void _chpl_comm_broadcast_global_vars(int numGlobals);

  This function is used to make sure that the wide classes on all
  locales other than locale 0 point to the class on locale 0.  A
  global array of pointers, char *_global_vars_registry[], contains a
  list of addresses on each locale.  (This array is generated by the
  compiler and initialized in compiler-generated code.)  On locale 0,
  these addresses point locally to a class; everywhere else they are
  uninitialized.  This function initializes them everywhere else with
  the addresses on locale 0.  This function is called collectively by
  each processor.  It is necessary to barrier when this function is
  called, though we can factor this barrier out too.

  void _chpl_comm_broadcast_private(void* addr, int size);

  This function is used to put the value assigned to a constant global
  scalar into that scalar on all of the other processors.  This
  function assumes that global variables have the same addresses on
  different processors.  This function should broadcast the value at
  addr onto all of the other processors.

Added primitives PRIMITIVE_HEAP_BROADCAST_GLOBAL_VARS and
PRIMITIVE_PRIVATE_BROADCAST to interface with these functions.  Added
primitive PRIMITIVE_HEAP_REGISTER_GLOBAL_VAR to interface with a macro
that puts global heap variables in the _global_vars_registry.

Added a call to a compiler-generated function heapAllocateGlobals from
within the runtime.  This function is used to initialize the
_global_vars_registry array and make a call to the runtime function
_chpl_comm_broadcast_global_vars.

Refactored buildOnStmt so that the fLocal flag is checked within this
function thus allowing it to be called within the compiler without the
guard.

Added functions _heapAllocGlobal and _heapAllocConstGlobal which work
similarly to _heapAlloc except that they are replaced later in
compilation so that global variables are allocated at program startup
instead of when their definition is executed.

The function _heapAllocConstGlobal handles constant globals and
specializes global variables of simple scalar type that are constants
by not allocating these variables on the heap.  Instead, when they are
defined, the definition is broadcast to all of the other locales.
This is an important optimization and is critical for handling certain
global variables, like maxThreads, so that it can be accessed in the
runtime code as a simple variable.

Added to the end of the normalization pass, a function
heapAllocateGlobals that searches for global variables and puts them
on the heap in a similar way that heapAllocateLocals worked for
variables accessed in begin statements.

Added code to insert temporaries to store nil when assigning nil to a
wide reference to a wide class.  Such assignments require nil to be
the size of a wide class containing nil, much like passing nil to a
function that requires a wide class.

Changed the semantics of .locale so that if .locale is applied to an
integral parameter, the parameter is taken to be a locale id.
Previously this would be an error, or should have been an error, since
parameters are not lvalues.  This allows us to use code like "on 0 use
..." for using the standard modules because the Locales array isn't
initialized until one of these modules is used.  This is slightly odd
because "on 0 do" has different semantics from "var i = 0; on i do"
since i is an lvalue and may exist on a locale other than 0.

Changed the semantics of a use-statement so that using a module
invokes an implicit "on Locales(0)" statement before calling that
module's initialization function.  This allows the
_run_module_firsttime variables to be private to a locale, but using
modules is still not thread safe.  Added a test case of using a module
from a remote locale.

Note that the _run_module_firsttime variables are initialized twice.
This is because they need to be initialized in a new function that is
called collectively by the runtime but generated by the compiler,
_initModuleGuards, since they are private to a processor.  They are
also initialized in the initialization of the _Program module.  This
is only done for function resolution because otherwise these variables
were considered to be uninitialized.  As a fix, it may be possible to
set their type and make them into compiler temporaries.

Moved the call of _startTrackingMem into _chpl_main so that it is
called after all of the standard modules are initialized and not from
within one of the module's initialization routines.  Because the
initialization functions are called within on-statements, we ran into
the situation where an alloc was called before tracking began but the
free called after tracking began.

Removed 5 futures relating to global variables and/or the interactions
between on- and begin-statements.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13151 3a8e244f-b0f2-452b-bcba-4c88e055c3ca